common: Use uintptr_t for storing pointer integer
Change pointer casting to uintptr_t to preseve the pointer and avoid
compiling warning on the architectures like CHERI where the pointer
is not an unsigned long integer.

- Use uintptr_t type for preseving original pointer.
- Use unsigned long type for extracting address from pointer.

+++ include/env.h
@@ -212,7 +212,7 @@ int env_set_hex(const char *varname, ulong value);
  */
 static inline int env_set_addr(const char *varname, const void *addr)
 {
-	return env_set_hex(varname, (ulong)addr);
+	return env_set_hex(varname, (ulong)(uintptr_t)addr);
 }
 
 /**
